store/app: add switchlanguage
