+++ webapp/store/app/actions/index.ts
@@ -1,10 +1,4 @@
-import axios from 'axios'
-
 import { Country, Region, RegionGroup } from '@core/country'
-import { createI18nPromise } from '@common/i18n/i18nFactory'
-import { applicationError } from '@webapp/components/error/actions'
-
-import * as UserState from '@webapp/store/user/state'
 
 import ActionTypes from './actionTypes'
 
@@ -23,19 +17,5 @@ export const updateRegionGroups = (regionGroups: Array<RegionGroup>) => ({
   regionGroups,
 })
 
-export const switchLanguage = (lang: any) => async (dispatch: any, getState: any) => {
-  try {
-    const userInfo = UserState.getUserInfo(getState())
-    if (userInfo) {
-      await axios.post(`/api/user/lang?lang=${lang}`)
-    }
-    const i18n = await createI18nPromise(lang)
-
-    if (lang === 'ar') document.body.classList.add('rtl')
-    if (lang !== 'ar') document.body.classList.remove('rtl')
-
-    dispatch({ type: ActionTypes.appI18nUpdate, i18n })
-  } catch (err) {
-    dispatch(applicationError(err))
-  }
-}
+export { initApp } from './initApp'
+export { switchLanguage } from './switchLanguage'
